Add support in Decompiler for float2x4 matrix.
Missing from before, and Lichdom needs it, so this adds that support.  Following the model of float4x2, roughly same size array.

Spot checked 4 shaders, all seemed right.

Added another fix for a nested 'if' that was all on one line, nearly certain that was a bug for non-array versions of float4x4, as it was all code on one line.
In any case float4x3 and float3x3 act differently now, worth noting.