Far Cry 4: Add convergence + weapon sight depth override while aiming
Right mouse button and left trigger on a controller override the
convergence and adjusts the depth of many weapon sights. This second
adjustment can affect other things in the game, so until we can use
texture or render target filtering it is only activated while aiming.

Also synchronise the d3dx.ini with the latest template.