Add support for type=cycle key bindings
This adds support for cycling values on a single key press. The syntax
looks like this:

[Key1]
Key = grave
Type = cycle
x = 0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.995
transition = 300,100

In this case the x list has 11 values, while the transition list only
has 2. The final elements of short lists are repeated, so this will
evaluate to (as printed in the log):

  Cycle 1: x=0 transition=300
  Cycle 2: x=0.1 transition=100
  Cycle 3: x=0.2 transition=100
  Cycle 4: x=0.3 transition=100
  Cycle 5: x=0.4 transition=100
  Cycle 6: x=0.5 transition=100
  Cycle 7: x=0.6 transition=100
  Cycle 8: x=0.7 transition=100
  Cycle 9: x=0.8 transition=100
  Cycle 10: x=0.9 transition=100
  Cycle 11: x=0.995 transition=100

Blank values are permitted in the list, which will omit that setting
from the cycle. A blank value can be used as the last value in a short
list to prevent it from being repeated, like this:

[Key2]
Key = left_control
type = cycle
x = 0, 0.25, 0.5, 0.75, 0.995
transition = 100,,

Which will evaluate to:

  Cycle 1: x=0 transition=100
  Cycle 2: x=0.25
  Cycle 3: x=0.5
  Cycle 4: x=0.75
  Cycle 5: x=0.995