Remove Key, TextureOverride, and ShaderOverride numbering & check for duplicate sections
[Key*], [TextureOverride*] and [ShaderOverride*] no longer have to end
in a number. This allows more deceptive free-form naming, such as
[KeyAim], which is especially useful for bindings that may be copied
from one config into another. Of course the old naming scheme will still
work as well, so existing configuration files do not need to be updated.

This will work even for section names of differing cases, like [KEY_FOO]
and [TeXtUrEoVeRRiDe1337] if someone was so inclined - this matches the
case insensitive nature of the existing ini parsing APIs.

Another benefit is that they no longer have to be consecutively numbered
- missing [Key2] will no longer prevent [Key3] from being parsed.

Duplicate sections (including those that differ only by case) are now
detected and will sound an audible beep to alert the user that they have
an error in their configuration. This is especially useful to detect
copy and paste errors.