Far Cry 4: Add shaders from Mike's WIP
These shaders have not been modified from Mike's WIP, except to remove
NULL bytes to ensure that git & github will treat them as text and show
diffs.

Mike noted that some of these shaders are likely broken as he had to
stop working on the fix at some point. I will also have to consolidate
differences between my WIP and Mike's in a future patch, to account for
differences in things like UI/crosshair depth and the repair tool flame.

These are the notes from Mike:

Last time I looked at the game (a few weeks ago) was just to see how the
game would run with two GTX 980's I got my hands on from work and I
noticed that the fire effects were wrong (like they are/were in non-SLI)
but I did not investigate.

I'm not sure if the MP3 profile will fix the antialiasing issue -
remember you need to use TXAA2 to prevent the 'frame delay' (which is a
visual killer for me).

The best thing to do is start a new game, go through the tedious 5
minutes movie and get to the part where they are sitting at the table.
You will see some wrong reflections on the table - I got the shader for
this but the 'standard' correction I used for other issues was not
fixing it, so there is something else that needs narrowing down. In that
room, when you turn round 180 degrees and face that far wall, I had
fixed all the lighting issues - if these look broken again you may be
suffering from me stopping mid-tweaking on something else (I literally
just had to drop this game at some point).

Other lighting issues you can see when you go down one level and there
are big fat pillars: some of the shadows when you get close look wrong -
this might be AO I'm not sure. I had a quick try of a couple of obvious
things, but again it needs a more creative look.

Finally, when you get past the driving section and you wake up next to
the dead driver, that area has a small stream which really highlights
the specular reflection issue that I was working on. For the life of me
I can't remember whether I was fixing the VS or PS or both, but I was
really close - there are a few different rendering issues you see, so as
you fix one the others break and vice versa, so it's like a puzzle to
shift them around in the right order lol.

I carried on after that area up until the hill top building where you
meet your buddies and start a fight, but the only other issues I saw
where a few HUD elements at screen depth.