Add copy_on_mark option, so that marking can be used in games where the decompiler crashes
Pirateguybrush reported crashes in Defence Grid 2 related to dumping
shaders as HLSL, either by export_hlsl or the copy on mark feature.

This option allows the later to be disabled, which can then allow the
hashes to be found in the log file and added to [ShaderOverride]
sections to disable them using that mechanism.