Refactor [ShaderOverride] and [TextureOverride] data structures
These sections were using separate maps for each of their settings, and
in the case of handling=skip was overloading the meaning of the
value in ShaderSeparationMap.

Refactor these to use a single map of data structures that store the
collection of settings supported by the sections, to make it easier to
add additional settings later.

Add an audible warning if multiple ShaderOverride or TextureOverride
sections have the same hash value as only one section will take effect.

Notably while refactoring this I have become convinced that the
iteration setting in both sections is broken and has been ever since
dropping DXGI support, and was further broken once the hunting buffers
were made to timeout. I will disable this code altogether in a separate
patch, until we have the means to make it work again in the future.