Disable broken [ShaderOverride] & [TextureOverride] iteration setting
The iteration setting depends on RunFrameActions() being called on a
frame boundary, which used to happen due to being called from present()
in DXGI.

These settings have been broken ever since dropping DXGI support and
using the draw() call to run frame actions.

It also appears that it was further broken by being associated with the
hunting buffers.

In the future we may be able to restore this if we either bring back
DXGI support or find another way to determine when the frame boundary
occurs.