Add convergence override option to [ShaderOverride] sections
This allows a custom convergence to be set while a shader is being used,
similar to the existing ability to set a custom separation value.

This can be used as a way to prevent effects from varying with
convergence as an alternate to modifying the stereo correction in the
shader - I'm adding it specifically to move the sky box in Defence Grid
2 back to infinity with convergence=0 since we are not yet able to dump
shaders from that game.