Add depth buffer filter
This is useful as a means to filter UI adjustments that also affect
other items in the world.

The depth buffer forms part of the heuristics that the driver uses to
determine whether to apply the stereo correction formula or not - if
there is no depth buffer, the formula is not applied (with some
exceptions - more experimentation needed), which is why UI elements are
usually displayed at screen depth regardless of their W coordinate.

At the moment this will use the replacement shader either when the depth
buffer is active (depth_fitler = depth_active), or when it is inactive
(depth_fitler = depth_inactive) and will use the unmodified shader
otherwise.

e.g. to prevent the weapon sight adjustment in Far Cry 4 from affecting
other objects in the world:

[ShaderOverrideWeaponSight]
Hash=765c3e296da52533
depth_filter = depth_inactive

Later it will probably be worthwhile adding an alternate mode that
passes whether the depth buffer is active or not into the shader so it
can use different code for either case, but we need to decide on the
best way to pass this information in first.

This patch also makes mCurrentVertexShader and mCurrentPixelShader valid
even while hunting mode is off so that the override settings can be
looked up in production mode - this has also fixed the handling=skip and
per-draw separation/convergence settings that were only working in
hunting mode.

If the performance impact of this change is too great we could look into
mapping the shader's address directly to the override structure instead
(care would need to be taken to ensure that the mapping would survive or
be reconstructed after a config reload).

This patch also has the result that marking_mode can be changed to
original via reload_config (marking_mode=zero still will not work).