Replace std::map with std::unordered_map where appropriate
std::map is sorted and has a lookup time of O(log N)

std::unsorted_map uses hash buckets to achieve an O(1) constant time
lookup in the average case, and a worst case lookup time of O(N) in the
case of severe hash collisions (unlikely)

Most of the maps we use do not need to be sorted, so change them to use
the faster unordered versions. This should alleviate some potential
performance issues in production mode.

Some maps & sets do need to be sorted - particularly any that are used
for the active shaders/buffers while hunting, any that print out
listings in the log on mark, and any that are used for the
ShaderUsage.txt so that it will be sorted. Notably these are all only
active in hunting mode. These have been noted in the globals.h file.