Use composition instead of multiple inheritance to clean up the input class hierarchy
The hierarchy got a little confusing when both XInput support and
repeating/delayed inputs were added around the same time, and both
inherited from InputAction for different purposes. This patch addresses
the issue by separating out the input backend from the action, which
removes the need for a number of boiler plate classes inheriting from
both.

An InputAction class or one of it's subclasses (RepeatingInputAction and
DelayedInputAction) are now constructed with an InputButton class, which
can be either a VKInputButton or an XInputButton.

Also merge the NewVKInputAction NewXInputAction factory functions into
the contractors with exceptions to handle parse errors.

This patch also fixes up some minor memory leaks that could occur when
reloading the config file - I had misunderstood what vector::clear()
would do when the vector held a pointer type.