Fix marking_mode=original and depth buffer filtering for shaders reloaded live
The original shader maps were only filled in with shaders that had been
replaced when first created, and not for shaders that were later
replaced via copy on mark or reload_shaders.

This prevented marking_mode=original and depth buffer filtering from
working on these shaders until the game was restarted.

To solve this add the original shader to the original shaders map
whenever hunting mode is enabled, at the same time the shader is
registered for possible reload.