Add show_original key binding to temporarily show the original game
Occasionally I run into a broken effect and find out that it was broken
by something else in the fix. marking_mode=original is a neat way to
check this, but can be pretty time consuming, and only shows one
original shader at a time.

This new key binding will temporarily use the original shaders from the
game as long as it is held, and will disable processing any
ShaderOverrides - this gives us a quick way to compare before and after
for the entire fix.