Add a 'partner' option to ShaderOverride to filter based on partner shaders
This can be used to make a fix to a vertex shader only apply when it is
used in conjunction with a specific pixel shader (or vice versa). If the
partner is matched the replacement will be used, otherwise the original
shader will be used.

At the moment only a single partner can be specified. Future work
includes allowing a list of shaders to be specified and a method to pass
in an identifier of the partner shader to the shader to allow more
advanced logic where a different fix may be needed for different
matching shaders.