Far Cry 4: Force weapon preview render targets to mono (FIXME: Resolution dependent!)
When previewing a weapon in the shop, the background is not cleared in
the left eye, which looks bad after previewing several weapons or
rotating one.

Force the 1920x1080 and 1280x720 versions to mono to avoid this issue.

NOTE: These render targets are 2x the screen resolution and as such will
not be special cased by the hashing function in 3Dmigoto, so this will
need to be fixed properly later (though for now 3Dmigoto does not have
access to the resolution so even the regular special case is broken).