Revert to original if replaced shader is deleted on reload_shaders
If a shader has been deleted from ShaderFixes reload_shaders will now
revert back to the original shader from the game. This means we no
longer have to restart the game if we want to remove a replaced shader,
as might happen if the decompiler has caused a rendering issue in a
shader we determine does not need to be fixed anyway.