Make traversing backwards past start of shaders wrap to the end
I keep missing shaders near the end of the list while hunting and can't
get back to them without cycling over the whole list again. This change
makes hunting wrap while going backwards as well so that it is easy to
quickly get back to a shader at the end of the list.