Add more Critical sections.
This pass was to go through EVERY set, map, and vector in Globals.h, and make sure that every assignment or clear or insert on those was wrapped with a critical section.

Probably minor problems, but numerous holes were patched here.

Some are unnecessary, I don't care.  I wanted to make it clear that each one is safe.