Refactor for new Hunting and IniHandler files.
These are loosely based on the DX10 experiment, and the goal was to pull all the hunting and iniparam stuff out of the main d3d11Wrapper file, leaving it's job to be just the wrapper.

This is not fully successful, because of some funny problems like RunFrameActions and EnableStereo, but the vast majority here make sense.

This compiles, but doesn't run yet.