Move EnableStereo to IniHandler.cpp and rename to CheckStereo.
The stereo init seems logically part of the rest of the init, including ini file parsing, so moved that there.

Have it be more robust on checking for stereo enabled, as when stereo is disabled we will crash later when trying to create a stereo texture.  Now it will beep and exit.