First pass at adding DrawString support as an Overlay.
This includes an Overlay object which calls out to the DirectXTK subproject, now included.  The reason to include the project instead of just a .lib for it, is to be able to step through the font code.

Includes some mods to the project files to get it to link.

Builds and runs in Alien though.