Rename files to make consistent file names.
This is using git mv, so that the file histories are preserved.

Changing the Direct3D11Device name to HackerDevice to reflect that it's all about the override.  Same for Context.

Making the names more closely reflect what they do.  Changing case on a few to make it start with Uppercase when they are cpp files.

non-Cpp files remain lowercase, like util, log.