Add GetParent overrides for HackerDXGIDevice and HackerDXGIAdapter.
These two are called through the secret d3d11 path to get access to the DXGI interface.  The override works, and allows the two objects to have their GetParent called, instead of the DXGIObject method.

Returning wrapped objects allowed it to get to the HackerDXGIFactory, which was also wrapped, and in fact allowed us access to the SwapChain.  So, success!

We now have access to the SwapChain for Mordor.