Add full wrapper for DXGI objects called from D3D11.
In CreateSwapChainAndDevice, we now wrap the SwapChain created to return a HackerDXGISwapChain instead, which will allow us to see Present calls.

Also add in HackerDXGIFactory::CreateSwapChain, the wrap to create a HackerDXGISwapChain so that it is wrapped there as well, if called directly.

This now works in Mordor to allow us to see and capture the Present calls.

This removes the use of HookedDXGI, as it's no longer necessary to hook the Present call when it's wrapped.  Wrapping is slightly better here because of that bizarre secret path for games that don't create a direct SwapChain.  We need to wrap all that, so wrapping the direct path here makes some sense.