Add ID3D11Device and ID3D11DeviceContext to HackerDXGI calls.
These are needed for the Overlay class, because we need the original Device and Context to be able to draw.
They are thus passed in when either the SwapChain for the Factory are created, and used during the SwapChain constructor to create the Overlay object.

Bit of a daisy chain of objects, might not be the best way to do this, but... it works!

In Mordor, for the first time ever, we have an overlay, and it's solid and doesn't flicker because it's drawn at Present().  That's an x64 game too.