Add the Overlay text for shader hunting.
These are VS and PS locations in the array of active shaders.
Shows N/X format to give an idea of where you are in the chain.

Not sure format or location is ideal, but it's a start.

Rewrote the Next/Previous PS and VS routines to be easier to read, and changed the use of the default variables to be -1 based when nothing is selected.
This was to fix an off by one error that using 0 caused, but was not previously visible.