Remove stereo enabled check.
Fixed a bug where it would crash if 3D was not enabled in GTA5, so this check seems to no longer be required.  Other people are using this for non-3D shader hacking, so this should allow them to be able to find and mark shaders.

Also fixed some x64 warnings in rewritten hunting code.

Works in GTA5 now, although full shader dumping crashes.