Update uninstall.bat to remove spritefont, leave d3dcompiler_46.
The d3dcompiler_46.dll will break the game if we delete it, and they need it, so for default builds here, we will now leave it.  If it's not used by the game, it doesn't hurt anything, and is better than breaking the games on occasion.

Add deletion of del courierbold.spritefont to the uninstall.bat, since we are dropping it in now too.

Delete  the courier.spritefont, as we are using bold version instead.