Basic structure for screen capture upon Mark.
The idea from Voodooman was to create a screen snapshot of the shader in question, tagged with the shader hash, so that we could use that as a reference for a given shader.

Seems like a good idea, although implementation is tricky.
Right now this will take a snapshot using DirectXTK, but it's not good enough.

In Mordor it makes a black snapshot, in Alien it makes a half-screen, not stereo.

But, this also brings the HackerDXGISwapChain reference into the corresponding HackerDevice, which should be useful.