Change default to export_hlsl=0 from 3.
The usual workflow for people does not include dumping all the shaders, so having this be 3 is a net negative because it takes a long time to start up, and it will crash on occasion, leading to confusion.

Setting to 0 by default to not dump shaders. More saavy users will know to find this for offline searches.