Add wrapping for DXGI EnumAdapters too.
Returns HackerDXGIAdapter now for each request.

Clean up the DirectXGI project to make it build under win32 debug at least.  Not quite ready, because of the use of Overlay which makes no sense in logging only case.

But tested it in Ori and did not see any other object leak.