Add overrides for QueryInterface to create IDXGIFactory2.
This handes the odd scenario where the caller calls HackerDXGIFactory::QueryInterface with riid=IDXGFactory2, looking to create a Factory2 object.

The DXGIWrapper is also changed for CreateDXGIFactory so that if they pass in riid=IDXGIFactory2, we create and return that as well.

This should handle the creation of Factory2, but I don't have a good way to test it.