Update IDXGIFactory::QueryInterface to return E_NOINTERFACE for Factory2 requests.
This is a bit of a hack, but it does workaround a crashing bug where anyone calling to create Factory2 through this back door would crash.

Factory2 is Win8.1 specific, and thus not mainstream just yet. The evil update on Win7 provides part of this functionality, but not all, and is optional. So it's not a giant leap to say we don't support Factory2 right now, as at least 50% of the machines in the field won't.

Tested to work in Ori, WoD, and Steam Overlay.