Fully straigthen out the HackerDXGI classes.
This removes all the duplicate code and duplicate methods that can just use their parent class methods. We now only make a subclass of a method when it's necessary to change behavior.

Fix one serious error of HackerDXGIOutput missing descending from the Hacker version.

This should be fully clean and structured now, and a good template for adding further objects we might need.

Tested and runs in Alien.