Change default on rasterizer_disable_scissor=1.
Was 0 by default for release builds, it seems like it makes more sense to default to disabling scissor limits as a way to avoid having to find and fix stuff.

Most general impact is slightly lower performance. Can in principle cause graphic glitches if the dev uses it as a way to clip HUDs or something.