Clean up the DXGI command handling.
There are numerous settings to force screen resolution, refresh, full screen, but they have been broken for awhile.

This cleans up all the usage of those, and makes sure that we are using the G-> versions.

Not sure that it's actually functional, but it compiles and runs and logs OK.