Tweak the UI for Mordor using new keybindings.
This is a start at improving the UI for the game, because the HUD is so busy.  These are shaders with good comments about their use and location.

Includes a d3dx.ini file that has some starting keys, including a cycle key to move the hud to different depths.

Game play UI is pretty solid, menu operations still need work.