Add Factory creation hooks using deviare in-proc.
This creates hooks for the CreateFactory and CreateFactory1 calls from the dxgi.dll.  This makes it possible for us to capture those calls, without having to wrap and ship a dxgi.dll.

The hooks are installed as early as possible during dllmain for d3d11.dll.

When called, it will create wrapped versions, which will then follow to other wrapped objects.

The goal of this sequence is get access to directly created CreateSwapChain results, which otherwise we would not wrap and not be able to Overlay.

This runs without crashing in pCars, but doesn't always draw overlay.