Skip outdated shader variants to fix crash in World of Diving
World of Diving (and presumably some other Unity games) contain shaders
that have both vs_4_0 and vs_2_x variants (or ps_4_0 and ps_2_x
variants). Unity appears to have some support for using DX11 at a DX9
feature level (the raw shaders in unity's asset files contain shaders
compiled as such), which this may be related to.

Having multiple shader models in the one shader confuses the decompiler
and leads to a crash due to an out of bounds access in applySwizzle(),
so teach the decompiler to ignore any instructions in a vs_2_* / ps_2_*
section of the shader.

This should be a reasonable solution as there is no real reason to use
the DX9 feature level (especially as Unity supports DX9 directly).

An alternative solution might be to treat both sections as separate
shaders, but that would take some effort for no clear benefit.