Add a new force_no_nvapi .ini parameter to force games 3D off.
This is the equivalent of editing the .exe and replacing nvapi with nvbpi to make the game think the library is missing.  Just easier, because it's just a parameter here.

Also tweak the automatic_mode setting so that it returns false for any calls to NvAPI_Stereo_IsEnabled to make it freerer for automatic mode to work.

Both of these techniques are tested and working in Project Cars.

The secret handshake between d3d11 and nvapi dlls is now changed to use a specific signature off of query_interface instead of something embedded in SetConvergence.