Remove the flag HOOK_SYSTEM32 as we now use hooking always.
This was originally expected to be game specific, but has proven to be generally useful, and now the hooking is used for DXGI support for the Overlay, which means we need it for x32 and x64 builds.

Removed from all build specifications, and conditional compilation.

This also fixes a bug introduced when nvapi64 was added to the list, for the x32 version of nvapi.  So both nvapi64.dll and nvapi.dll are now bypassed back to the game directory.