Simplify the LoadLibraryExW hooking.
There are three uses cases we need to handle, and this cleans up all the paths to the simplest.
It also adds more error checking on calls that can theoretically fail, and adds a DoubleBeepExit handler to handle fatal errors, rather than crash the game.

I don't have a way to test the 3rd case of 32 bit game on 32 bit OS, but it's consistent with the other two.

Works in Ori and Project Cars.  x32 and x64.