Witcher 3: Fix decompiler issue in shadow shader due to int/float mixup
Helifax reported this shadow shader did not work correctly in caves,
appearing to be in sunlight instead of shadow.

Turns out it is due to reading integers from the constant buffers as
floats (which is of course due to the missing headers leading us to
treat everything as floats). Add asint() reinterpret casts around each
of the affected loads to fix the issue.