Add TextureCubeArray handling and fix dcl_textures for non-header case.
In the non-header info case, the dcl_texture_* types need to be sized.  Unfortunately all size info has been stripped, including text headers, so we can't determine the proper size, and at the moment just default to 4. This will generate some hand-fix problems, but I don't see any solution.

Also add TextureCubeArray handling as a new variant that matches the others.  This was working in the header case, but missing in the no-header case.