Fix every I**/U** type instruction to use asint() and asuint().
We were doing direct casts before like (int)x, and that creates type coercion code, which is not what we want for integer instructions.  For integer instructions we want the contents to be treated as int, with no conversion, which is what the asint() intrinsic does.

So, switched every use of coercion to be asint(x) and asuint(x) instead.

Only done on instructions that are explicitly integer, like IADD, ILT, ULT, INEG, USHR, and so on.

This should allow the fxc compiler to determine the proper instruction, and avoid the extra ftoi and utof type instructions we've been generating.

This has not been well tested, but it did not seem any worse than before on Witcher, and spot checks looked right.