Witcher 3: Redo some fixes to check decompiler improvements
This redoes four of the shaders affected by the constant buffer
float/int misinterpretation bug, and confirms that the new behaviour
seems good. show_original identified some small rendering issues in some
of these, which are now resolved :)

The most notable shader not redone in this commit is
7c44a1f4840ef159-ps_replace.txt (e.g. light from fireplace in the
starting room), as it has additional unresolved decompiler issues
causing it to lose shadows.