Partial revert of "Remove the export_binary option from d3dx.ini"
I want to be able to extract the raw binary shaders for examination and
comparisons with the versions produced by the assembler, so add this
option back in, but leave it hidden by not documenting it in the
d3dx.ini since very few people would ever want it.

Code was slightly refactored.