Add utility routines to Disassemble and write text shaders.
Lots of copy pastes of these code sequences, goal here is to make one routine that can be used in all those cases.

One new variant creates a std::string for a given disassemble, which avoids the problem of the trailing null byte.
The other writes the actual text file, so we have spot if we decide to change the naming.

Only does one location right now for export_shaders.