Use utility routines to handle most of the file creation.
For Asm files and asm conversions, this now uses the new utility routines to put that all in one spot.

Hunting uses these now instead of having a copy, as well as ReplaceShader.

Simplified numerous sections to make them more readable.

Assembler succeeds on an unmodified shader, but fails if we edit it.