Fix crashing bug if 'evil' update is not installed.
The 'evil update' is the windows platform update that brings DXGI 1.2 to Win7.  We were crashing when that update was NOT installed, which was the opposite of what we'd expect.

This fixes that bug which was caused by passing a HackerDevice to the real CreateSwapChain.  Not surprisingly, it didn't care for it.  Surprisingly, with platform update installed, that would still work.

Tested in pCars and Witcher3, and fixes crashes in both.

Need to test with update re-installed.