Beef up the logging to show object pointers too.
There object names are helpful, but for Dragon Age I wanted to see if it was the same object or not, so added in a %s@%p format to see name@pointer values.

Clean up some other logging, add consistency to it.

Add a few more mystery UUIDs to name conversion routine.