Fix not keeping original shaders for show_original in some circumstances
Add the missing condition on show_original being enabled in the config
file as one possible reason to keep original shaders around (I was
probably relying on the fact I always turn on reload_config).

Since hunting may be soft enabled now remove it as one of the
conditions. Later if we distinguish between hard & soft enabled then
only soft enabled should allow original shaders to be kept if they are
not required for a filter.