Return E_NOINTERFACE for any QueryInterface upcast to ID3D11Device1.
Dragon Age showed this problem, by upcasting to Device1 they escaped the net, and could create a device we didn't wrap, which ends in a crash.

Added wrappers for all that, but it occurred to me that it might make more sense to just return an error here, since this is an optional interface for 50% of the users.

Unless or until we find some negative result from this, going to a simpler and more reliable API is better for our stability.

Fixes the crash in DragonAge.