Fix crashing bug in CreateDeferredContext(), and re-enable wrapping
The crash was due to incorrectly checking the pointer pointed to by
ppDeferredContext for NULL rather than the pointer to the pointer itself
as the condition to execute the code or not. We therefore always skipped
the code (unless the application itself had not bothered to properly
initialise the pointer to NULL) and returned a NULL pointer, leading to
the application to crash when it dereferenced it.

Was hoping this would help identify a shader in The Witcher 3 that I
suspected might be rendered from another thread... It didn't get me any
closer, but still seems like a worthwhile change.