Tweak overlay to be centered on top and bottom.
Didn't really like the left hand nature of overlay, so this is a small tweak to try out having it be centered with shader hunting on top of screen, sep+conv on bottom.  So split them up.

Initial look seems pretty good.

This will scale to proper screen dimensions, although the font is fixed size and appear to shrink on bigger resolutions.