Fix ref counting bugs in GetImmediateContext()
GetImmediateContext() is supposed to increase the device ref count, so
the application is expected to call Release() when done. By not calling
the original function the ref count would be one lower than expected,
leading to a possible premature free as the application is shutting
down (crash on exit type symptom).