Straighen out the QueryInterface and GetImmediateContext calls.
These are getting fairly verbose now, with lots of different variants.  Trying to create a more logical flow here, with the primary change being that we do the original call first so that ref counting is working correctly, and we have the objects necessary.

Adds IDXGIDevice1 handler in QueryInterface, as WatchDogs needs that.
Add IDXGIDevice2 handler to return an error to follow our platform/evil update philosophy.

Handle possible nullptr errors passed in to CreateDeferredContext and Context1.

Handle possible out of sequence creation of mHackerContext if CreateDeferredContext gets called too early.  And log all that jazz so we can see if we need to tune it.