Make HackerDevice constructor minimalist, move init to CreateStereoAndIniTextures.
The constructor was doing a lot of work, including initializing the stereo handle and iniparams handle.

That was problematic because the stereo creation would call back to the device for ImmediateContext, and that would not yet be setup.

Moved all of that code into a creator, CreateStereoAndIniTextures that happens after the objects are setup and linked to each other.

This solves the problem of arriving at GetImmediateContext with a null mHackerContext.