Put overlay into it's own context to save & restore state (unworkable)
This approach doesn't work with single threaded games as deferred
contexts cannot be created. Forcing these games to enable multi-threaded
rendering on device creation leads to a crash instead.

It also doesn't solve the pipeline errors (though the error has
changed), but there's not much point pursuing this further given the
above.

Leaving this attempt in the git history in case we want to revisit it
again in the future.