Fix crash if CreateSwapChain is called with a DXGIDevice instead of a D3D11Device
UE4 calls CreateSwapChain() using a DXGIDevice instead of the
D3D11Device that we were expecting. This is perfectly legal as
CreateSwapChain takes an IUnknown, so we need to deal with it.

This also raises the possibility that other devices (e.g. D3D11Device1,
D3D10Device, D3D12Device, etc) may be used with CreateSwapChain, which
we don't handle at present.