Move the RunFrameActions into DXGI.
This is the logical place for it, called off of the Present() call. Having it in Context was a hack to avoid needing DXGI, but now that DXGI is working OK, it's time to return it.

This also requires having the HackerDXGISwapChain know about the HackerDevice as an instantiation time reference.