Fix an out of sequence use of mStereoHandle in Overlay.
Because we need to do late initialization on the HackerDevice itself, that means we can create the overlay before the HackerDevice is fully set up.

Rather than save the mStereoHandle reference, we'll just save the HackerDevice reference as it will be set up by the time we actually need to use it.