Small tweak to CreateSwapChain to use original device.
We inadvertenly changed the object type passed into the master CreateSwapChain from an IDXGIDevice into an ID3D11Device instead.

Keeping with the philosophy of trying to not change the original calls, this seemed like it might be a problem.

Haven't seen any side effect, but it seems more reliable to use whatever device the game passed in originally.