Refactor HackerDXGI again to reflect new info.
Since we've learned more about how DXGI is used by games, this refactor simplifies the objects, and uses a hopefully more reliable approach for code style.

We now call the original call with the same parameters passed in, in every case.  And restructure them to have a Succeed clause instead of early exit.

Remove almost all of the duplicate mHackerDevice/Context references, as the only ones that matter are for HackerDXGISwapChain.

Be more careful about possible null pointer references from game bugs. May not but fully complete, but the overall structure should be solid.

Remove GetParent from the two subclasses, and put it back in HackerDXGIObject as the base class for that.

Clean up the header declarations to match the original D3D11 headers, with no 'hacker' variants as inputs. We now use reinterpret_cast instead.

Tested and runs on Witcher3 and Caffeine Demo as a UE4 example.